Redoing the file structure for the parse.
