+++ README.md
@@ -1,19 +1,17 @@
-# TASKS:
+## Tasks List:
 
-- Code the tokenizer (Lexer).
-- Create the AST using the output of the lexer 
-- Validate the syntax using the lexer.
-- syntax validation
-- variables extraction
-building a tokenizer
-haj gigachad        |         haj matich
-________________________________________________________________
-lexer               |   build AST using recusive descent
-variable expansion  |   shell history
-pipes               |   exit
-redirictions        |   export 
-AST traversing      |   signale handling
-execution           |   $? expansion
-echo                |   unset
-cd                  |   env
-pwd                 |    
+The following list contains the tasks that should be performed by collaborators:
+
+**Yassir58:**
+- [x] Code the tokenizer (**Lexer**)
+- [x] Create a syntax validator
+- [ ] Create an execution chain
+- [ ] Handle the redirections
+- [ ] Handle the pipes
+- [x] Builtins (**cd echo pwd**)
+
+**lhaj-matich:**
+- [x]  Create execution nodes (**Parse**)
+- [ ]  Shell history
+- [ ] Builtins (**unset exit export env**)
+- [ ] $? Expansion
